Fix Jest: split OpenAPI ref helpers from template-loader (#156 tests)
- Add openapi-schema-refs.ts for extractClassNameFromRef / findReferencedClasses
- group-schema-export imports refs module only so suite loads without top-level await
- Re-export from openapi.ts for existing imports; bump objectified-ui patch

Made-with: Cursor
