+++ objectified-ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "objectified-ui",
-  "version": "0.8.58",
+  "version": "0.8.59",
   "private": true,
   "scripts": {
     "dev": "next dev",
